Move all UI-side RPC wrappers into herder_facade
