+++ src/herder_facade/mod.rs
@@ -1,0 +1,6 @@
+//! Utilities for spawning and interacting with herder daemons.
+
+mod evdist;
+mod herder;
+
+pub use herder::{Herder, StartWriterError, WriterHandle};
