+++ src/main.rs
@@ -5,10 +5,9 @@ mod byteseries;
 mod compression;
 mod device;
 mod escalation;
-mod evdist;
 mod hash;
 mod hashfile;
-mod herder;
+mod herder_facade;
 mod herder_daemon;
 mod ipc_common;
 mod logging;
